Fix Diesel Generator burning fuel for too short of time, closes #6203 (#6204)
Fixes preemptive cast to int when calculating how long the generator
should run on small amounts of fuel.

If the generator did not consume the full amount of `toConsume` then
`state.consumeTick` would be truncated to 0, causing the generator to
consume fuel again on the next tick.

Now the generator will calculate the correct amount of ticks to run for
given a partial amount of fuel.

Fixes #6203